requirements.txt
fastapi
uvicorn
stripe
python-dotenv
mysql-connector-python
pydantic
jinja2
email-validator
